feat: admin auth

+++ pages/api/v1/achievement_post/[id]/delete.ts
@@ -1,9 +1,14 @@
 import { PrismaClient } from '@prisma/client';
 import type { NextApiRequest, NextApiResponse } from 'next';
+import { isAdmin } from 'utils/api/isAdmin';
 
 const prisma = new PrismaClient();
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+  if (!(await isAdmin(req, res))) {
+    return res.status(401).end();
+  }
+
   const id = Number(req.query.id);
 
   await prisma.achievementPost.delete({
